added address entity also to contact in completion form
