+++ Controller/AbstractController.php
@@ -13,6 +13,10 @@ namespace Sulu\Bundle\CommunityBundle\Controller;
 
 use Sulu\Bundle\CommunityBundle\DependencyInjection\Configuration;
 use Sulu\Bundle\CommunityBundle\Manager\CommunityManagerInterface;
+use Sulu\Bundle\ContactBundle\Entity\Address;
+use Sulu\Bundle\ContactBundle\Entity\AddressType;
+use Sulu\Bundle\ContactBundle\Entity\ContactAddress;
+use Sulu\Bundle\ContactBundle\Entity\Note;
 use Sulu\Bundle\SecurityBundle\Entity\User;
 use Symfony\Bundle\FrameworkBundle\Controller\Controller;
 use Symfony\Component\Form\FormInterface;
@@ -23,11 +27,6 @@ use Symfony\Component\HttpFoundation\Response;
  */
 abstract class AbstractController extends Controller
 {
-    /**
-     * @var CommunityManagerInterface[]
-     */
-    private $communityManagers;
-
     /**
      * @var string
      */
@@ -160,5 +159,61 @@ abstract class AbstractController extends Controller
     private function getTemplateAttributes($custom = [])
     {
         return $this->get('sulu_website.resolver.template_attribute')->resolve($custom);
+    }
+
+    /**
+     * {@inheritdoc}
+     *
+     * @return User
+     */
+    public function getUser()
+    {
+        /** @var User $user */
+        $user = parent::getUser();
+
+        if (null === $user->getContact()->getMainAddress()) {
+            $this->addAddress($user);
+        }
+
+        if (0 === count($user->getContact()->getNotes())) {
+            $this->addNote($user);
+        }
+
+        return $user;
+    }
+
+    /**
+     * Add address to user.
+     *
+     * @param User $user
+     */
+    private function addAddress(User $user)
+    {
+        $entityManager = $this->get('doctrine.orm.entity_manager');
+        $contact = $user->getContact();
+
+        $address = new Address();
+        $address->setPrimaryAddress(true);
+        $address->setNote('');
+        $address->setAddressType($entityManager->getRepository(AddressType::class)->find(1));
+        $contactAddress = new ContactAddress();
+        $contactAddress->setAddress($address);
+        $contactAddress->setContact($contact);
+
+        $contact->addContactAddress($contactAddress);
+    }
+
+    /**
+     * Add note to user.
+     *
+     * @param User $user
+     */
+    private function addNote(User $user)
+    {
+        $note = new Note();
+        $note->setValue('');
+        $user->getContact()->addNote($note);
+
+        $this->get('doctrine.orm.entity_manager')->persist($note);
     }
 }
