+++ Controller/ProfileController.php
@@ -86,58 +86,4 @@ class ProfileController extends AbstractController
             ]
         );
     }
-
-    /**
-     * {@inheritdoc}
-     *
-     * @return User
-     */
-    public function getUser()
-    {
-        /** @var User $user */
-        $user = parent::getUser();
-
-        if (null === $user->getContact()->getMainAddress()) {
-            $this->addAddress($user);
-        }
-
-        if (0 === count($user->getContact()->getNotes())) {
-            $this->addNote($user);
-        }
-
-        return $user;
-    }
-
-    /**
-     * Add address to user.
-     *
-     * @param User $user
-     */
-    private function addAddress(User $user)
-    {
-        $entityManager = $this->get('doctrine.orm.entity_manager');
-
-        $address = new Address();
-        $address->setPrimaryAddress(true);
-        $address->setAddressType($entityManager->getRepository(AddressType::class)->find(1));
-        $contactAddress = new ContactAddress();
-        $contactAddress->setAddress($address);
-        $contactAddress->setContact($user->getContact());
-
-        $user->getContact()->addContactAddress($contactAddress);
-    }
-
-    /**
-     * Add note to user.
-     *
-     * @param User $user
-     */
-    private function addNote(User $user)
-    {
-        $note = new Note();
-        $note->setValue('');
-        $user->getContact()->addNote($note);
-
-        $this->get('doctrine.orm.entity_manager')->persist($note);
-    }
 }
